% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivNormACHO.R
\name{bivNormACHO}
\alias{bivNormACHO}
\title{The ACHO function for given distributions (2D)}
\usage{
bivNormACHO(mu1, sig1, mu2, sig2)
}
\arguments{
\item{mu1}{mean vector of the first set.}

\item{sig1}{covariance matrix for the first set.}

\item{mu2}{mean vector of the second set.}

\item{sig2}{covariance matrix for the second set.}
}
\value{
The general "ACHO" value of the normal distributions.
}
\description{
This function calculates a value similar to the "Averaged Convex Hull Overlap" based on the prediction ellipses of the given normal distributions.
}
\examples{
library("mvtnorm")
n <- 300
s1_x <- 2
s2_x <- 4
rho_x<- 3/4
m_x <- c(0,0)
s_x <- matrix(c(s1_x^2,rho_x*s1_x*s2_x,rho_x*s1_x*s2_x,s2_x^2), ncol=2)
X <- rmvnorm(n, m_x, s_x, method="chol")

s1_y <- 2
s2_y <- 4
rho_y<- -3/4
m_y <- c(2,3)
s_y <- matrix(c(s1_y^2,rho_y*s1_y*s2_y,rho_y*s1_y*s2_y,s2_y^2), ncol=2)
Y <- rmvnorm(n, m_y, s_y, method="chol")

plot(X, main="", xlab="", ylab="", xlim=c(-8,8), ylim=c(-15,15), type="n", bty="n")
abline(h=0, v=0)
points(X, col="blue", pch=20)
points(Y, col="red", pch=20)

bivNormACHO(m_x, s_x, m_y, s_y)
}
