% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawCHull.R
\name{drawCHull}
\alias{drawCHull}
\title{Draws convex hull (2D)}
\usage{
drawCHull(data, color = "blue", line_width = 1, line_type = 1)
}
\arguments{
\item{data}{input data.}

\item{color}{color of the convex hull lines.}

\item{line_width}{line width of the convex hull.}

\item{line_type}{line type of the convex hull.}
}
\value{
None (convex hull polygon).
}
\description{
This function draws the convex hull for the input data onto an existing plot.
}
\examples{
n <- 100
X <- cbind(rnorm(n, 0, 2),rnorm(n,0,1))

plot(X, xlab="", ylab="", pch=20, bty="n")
abline(h=0, v=0)
drawCHull(X)

plot(X, xlab="", ylab="", pch=20, bty="n")
abline(h=0, v=0)
drawCHull(X, color="red", line_width=2, line_type=2)
}
