% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawPredEllipse.R
\name{drawPredEllipse}
\alias{drawPredEllipse}
\title{Draws prediction ellipses (2D)}
\usage{
drawPredEllipse(
  mu,
  sig,
  alphas = c(0.1, 0.5, 0.99),
  color = "blue",
  line_width = 2,
  line_type = 2
)
}
\arguments{
\item{mu}{mean vector of the normal distribution.}

\item{sig}{covariance matrix of the normal distribution.}

\item{alphas}{quantiles of the prediction ellipses to be drawn.}

\item{color}{color of the prediction ellipses.}

\item{line_width}{line width of the prediction ellipses.}

\item{line_type}{line type of the prediction ellipses.}
}
\value{
None (prediction ellipses polygon).
}
\description{
This function draws the prediction ellipses for the input data onto an existing plot.
}
\examples{
library("mvtnorm")
s1 <- 2
s2 <- 4
rho<- 3/4

n <- 500
m <- c(0,0)
s <- matrix(c(s1^2,rho*s1*s2,rho*s1*s2,s2^2), ncol=2)
X <- rmvnorm(n, m, s, method="chol")

plot(X, main="", xlab="", ylab="", xlim=c(-10,10), ylim=c(-15,15), type="n")
abline(h=-3:3*5, v=-3:3*5, col="grey", lty=2)
abline(h=0, v=0)
points(X, col="green", pch=20)

drawPredEllipse(mu=m, sig=s, alphas=c(0.5,0.75), color="blue")
drawPredEllipse(mu=m, sig=s, alphas=c(0.25,0.9), color="red", line_width=1, line_type=1)
}
