% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAcho.R
\name{getACHO}
\alias{getACHO}
\title{The base ACHO function}
\usage{
getACHO(data1, data2, acho_plot = FALSE)
}
\arguments{
\item{data1}{first dataset (1D/2D).}

\item{data2}{second dataset (1D/2D).}

\item{acho_plot}{Determines if the convex hull overlap behaviour plot is shown. Default TRUE.}
}
\value{
The "ACHO" value of the datasets.
}
\description{
This function calculates the "Averaged Convex Hull Overlap" for the input and can also show a plot for the behaviour of the reducing overlaps.
}
\examples{
n <- 1000
x1 <- rnorm(n, 0, 1)
x2 <- rnorm(n, 1, 1.5)

plot(density(x1), main="", xlab="", ylab="", type="n", bty="n")
lines(density(x1), col="blue")
lines(density(x2), col="red")

getACHO(x1, x2)
getACHO(x1, x2, acho_plot=TRUE)
}
