% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normACHO.R
\name{normACHO}
\alias{normACHO}
\title{The ACHO function for given distributions (1D)}
\usage{
normACHO(mu1, sig1, mu2, sig2)
}
\arguments{
\item{mu1}{mean vector of the first set.}

\item{sig1}{standard deviation for the first set.}

\item{mu2}{mean vector of the second set.}

\item{sig2}{standard deviation for the second set.}
}
\value{
The general "ACHO" value of the normal distributions.
}
\description{
This function calculates a value similar to the "Averaged Convex Hull Overlap" based on the centralized quantiles of the given normal distributions.
}
\examples{
n <- 1000
s <- seq(-5, 6, by=0.01)

m1 <- 0
s1 <- 1
x1 <- dnorm(s, m1, s1)
m2 <- 1
s2 <- 1.5
x2 <- dnorm(s, m2, s2)

plot(s, x1, xlab="", ylab="", type="n", bty="n")
lines(s, x1, col="blue")
lines(s, x2, col="red")

normACHO(m1, s1, m2, s2)
}
