% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCHull.R
\name{plotCHull}
\alias{plotCHull}
\title{Plots convex hull (2D)}
\usage{
plotCHull(
  data,
  color = "blue",
  points_color = "black",
  line_width = 1,
  line_type = 1
)
}
\arguments{
\item{data}{input data.}

\item{color}{color of the convex hull lines.}

\item{points_color}{color of the data points.}

\item{line_width}{line width of the convex hull.}

\item{line_type}{line type of the convex hull.}
}
\value{
Plot with data points and their convex hull.
}
\description{
This function plots the data and the convex hull for it.
}
\examples{
n <- 100
X <- cbind(rnorm(n, 0, 2),rnorm(n,0,1))

plotCHull(X)
plotCHull(X, color="red", points_color="blue", line_width=2, line_type=2)
}
