% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredEllipse.R
\name{plotPredEllipse}
\alias{plotPredEllipse}
\title{Plot prediction ellipses (2D)}
\usage{
plotPredEllipse(
  mu,
  sig,
  alphas = c(0.1, 0.5, 0.99),
  color = "blue",
  points_color = "black",
  line_width = 2,
  line_type = 2
)
}
\arguments{
\item{mu}{mean vector of the normal distribution.}

\item{sig}{covariance matrix of the normal distribution.}

\item{alphas}{quantiles of the prediction ellipses to be drawn.}

\item{color}{color of the prediction ellipses.}

\item{points_color}{color of the data points.}

\item{line_width}{line width of the prediction ellipses.}

\item{line_type}{line type of the prediction ellipses.}
}
\value{
Plot with data points and their prediction ellipses.
}
\description{
This function plots the data and the prediction ellipses for it.
}
\examples{
s1 <- 2
s2 <- 4
rho<- 3/4

m <- c(3,5)
s <- matrix(c(s1^2,rho*s1*s2,rho*s1*s2,s2^2), ncol=2)

plotPredEllipse(mu=m, sig=s)
plotPredEllipse(mu=m, sig=s, alphas=c(0.25,0.5,0.75,0.9), color="red", line_width=1, line_type=1)
}
